# terraform/modules/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: SkillForge AI API
  description: |
    SkillForge AI platform API providing access to all microservices through a unified gateway.
    
    ## Authentication
    This API supports multiple authentication methods:
    - JWT Bearer tokens for user authentication
    - API Keys for service-to-service communication
    - IAP (Identity-Aware Proxy) for additional security layer
    
    ## Rate Limiting
    API calls are rate limited to ensure fair usage and system stability.
    
    ## CORS
    Cross-Origin Resource Sharing (CORS) is enabled for web applications.
  version: 1.0.0
  contact:
    name: SkillForge AI Team
    email: api@skillforge.ai
  license:
    name: Proprietary
    url: https://skillforge.ai/license
servers:
  - url: https://api.skillforge.ai
    description: Production server
  - url: https://api-staging.skillforge.ai
    description: Staging server

# Security Schemes
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
    
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: API key in query parameter
    
    IAPAuth:
      type: http
      scheme: bearer
      description: Identity-Aware Proxy authentication
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Authentication required"
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Rate limit exceeded"
              message:
                type: string
                example: "Too many requests"
              retry_after:
                type: integer
                example: 60

# Global security requirements
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - ApiKeyQuery: []

# API Paths
paths:
...
  /api/v1/companies:
    get:
      summary: List companies
      description: Retrieve a list of companies
      operationId: listCompanies
      tags:
        - Companies
      responses:
        '200':
          description: List of companies
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-google-backend:
        address: "${company_service_url}"
    
    post:
      summary: Create company
      description: Create a new company
      operationId: createCompany
      tags:
        - Companies
      responses:
        '201':
          description: Company created successfully
        '400':
          description: Invalid input
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-google-backend:
        address: "${company_service_url}"
  /api/v1/companies/{companyId}:
    get:
      summary: Get company by ID
      description: Retrieve a specific company by ID
      operationId: getCompanyById
      tags:
        - Companies
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company details
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Company not found
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-google-backend:
        address: "${company_service_url}"

  # Subscription Service Routes
...
  /api/v1/gamification/points:
    get:
      summary: Get user points
      description: Retrieve user's points and ranking
      operationId: getUserPoints
      tags:
        - Gamification
      responses:
        '200':
          description: User points and ranking
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-google-backend:
        address: "${gamification_service_url}"
...